update CHP definition
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1147411 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hwpf/usermodel/CharacterRun.java
@@ -397,7 +397,7 @@ public final class CharacterRun
 
   public void setVerticalOffset(int hpsPos)
   {
-    _props.setHpsPos(hpsPos);
+    _props.setHpsPos((short) hpsPos);
     _chpx.updateSprm(SPRM_HPSPOS, (byte)hpsPos);
   }
